backend/requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
Flask-Cors==4.0.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
eventlet==0.33.3
